pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-trace"
version = "0.1.0"
description = "Tracing middleware for FastMCP with pluggable adapters (local file, PostgreSQL, OpenTelemetry, etc)."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
license = { file = "LICENSE" }
dependencies = [
    "psycopg2-binary",
    "requests",
    "pydantic",
    "fastmcp"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.setuptools.packages.find]
include = ["mcp_trace*"] 